* origin/dizzy: (95 commits)
  Update VirtualBox kernel modules
  Check if VirtualBox kernel modules do exist
  toradex: move to V2.4 branch
  qt5: use 5.5 branch
  Change all tab characters to spaces
  boundary: include uvcvideo kernel module
  Add Emulator sdcardmountd
  Include tzdata and ntp tools to image
  qt5: add declarative tools and host libs to sdk
  Some projects were moved from internal gerrit to codereview
  Use correct versions for addons
  Use correct path for QT_SDK_PATH
  fonts: update checksums for otf-noto
  gstreamer: make gstreamer version a DISTRO_FEATURE
  b2qt-demos: add QtQuick Controls demos
  Increase CMA reserved size on nitrogen6x.
  qtquickcompiler: disable for BYOS
  Remove recipe for QtQuick Enteprise Controls
  Support BYOS for enterprise only addons
  b2qt-demos: include qtcanvas3d planets demo
  ...

Conflicts:
	recipes/hunspell/hunspell_1.3.2.bb

+++ recipes/opengldummy/files/headers/EGL/eglplatform.h
@@ -105,9 +105,9 @@ typedef void*                           EGLNativeDisplayType;
 #if 1
 
 /* QtGlesStream */
-typedef unsigned int EGLNativeDisplayType;
-typedef unsigned int EGLNativeWindowType;
-typedef unsigned int EGLNativePixmapType;
+typedef intptr_t EGLNativeDisplayType;
+typedef intptr_t EGLNativeWindowType;
+typedef intptr_t EGLNativePixmapType;
 
 #else
